requirements.txt
pip==18.1
Django==2.0.9
djangorestframework==3.8.2
PyMySQL==0.8.1
pysaml2==4.6.3
django-model-utils==3.1.2
django-appconf==1.0.2